🔨 Refactor: 변수명 및 컴포넌트 리팩토링
